+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0-alpha.18] - 2026-09-06
+
 ### Added
 - **Startup & Periodic Auto-Update Checks**: Check for updates immediately on daemon startup and periodically every 6 hours.